combined class and function Stage tests. Removed internal hook export.
